requirements.txt
gitpython==2.1.15
python-dotenv==0.11.0
python-telegram-bot==12.4.2
